+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - Make `colpair_map()` more robust to input column names, with the exception of ".data" and ".env" (@jameslairdsmith, #131).
 
+- `correlate()` now removes non-numeric columns from data frame inputs (@thisisdaryn, #139).
+
 # corrr 0.4.3
 
 - Fix EOL issues and class attribute (@krlmlr, #93 and #90)